requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==3.2.19
djongo==1.3.6
dnspython==2.3.0
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
kombu==5.2.4
mongoengine==0.27.0
oauth2client==4.1.3
prompt-toolkit==3.0.38
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==3.12.3
pyparsing==3.0.9
python-snappy==0.6.1
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.6.2
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
winkerberos==0.9.1
zipp==3.15.0
